Simplify pattern-matching expressions
Initially, the grammar's regular expressions were "optimised" to be more
thoughtful in how they matched against keywords. For example, instead of
testing several similar-looking strings individually, a nested capturing
group (or class) could be used instead:

    Simple:    3D_Studio|3D_Studio_Shape|x_axis|y_axis|z_axis
    Nested:    3D_Studio(?:_Shape)?|[xyz]_axis

This was done in an effort to reduce overhead, as it was believed having
more substrings to run through would require lengthier processing time.

Bench-testing this revealed something altogether different, though. When
run 10 times on a repo of non-trivial size (145.61 MBs / 500 files), the
"simpler" expressions actually proved to be marginally faster than their
"optimised" counterparts on average:

            simple      nested

    real    27.8927s    28.0916s
    user    27.7411s    27.9443s
    sys     0.1439s     0.1405s

While this is a negligible performance gain, it does demonstrate there's
no point of a developer going through the added effort of hand-compiling
the expressions to save overhead. This approach is also more sustainable
in long-term, as well: no developer would prefer picking through complex
behemoths of nested capturing groups in order to remove a certain term.